requirements.txt
fastapi==0.95.1       # FastAPI framework
uvicorn==0.22.0       # ASGI server (for local testing)
mangum==0.17.0        # AWS Lambda + API Gateway adapter for FastAPI
boto3==1.26.0         # AWS SDK for Python (to interact with DynamoDB, S3, Cognito, etc.)
pydantic==1.10.7      # Data validation & serialization
pyjwt==2.6.0  # For decoding JWT tokens